@@ -2,26 +2,67 @@
   "name": "sass-scoped-media-query",
   "version": "1.0.0",
   "description": "Scope CSS styles and apply them only inside provided selector and media query pairs.",
-  "main": "_index.scss",
-  "author": "Ivan Nikolić <niksy5@gmail.com> (http://ivannikolic.com)",
   "license": "MIT",
+  "author": "Ivan Nikolić <niksy5@gmail.com> (http://ivannikolic.com)",
+  "sideEffects": false,
+  "exports": {
+    ".": {
+      "sass": "_index.scss",
+      "style": "_index.scss",
+      "default": "_index.scss"
+    },
+    "./package.json": "./package.json"
+  },
+  "main": "_index.scss",
   "style": "_index.scss",
   "directories": {
     "test": "test"
   },
+  "files": [
+    "_index.scss",
+    "lib/",
+    "CHANGELOG.md",
+    "LICENSE.md",
+    "README.md"
+  ],
   "scripts": {
-    "test": "eslint {test/**/*}.js && stylelint {_index,test/**/*.scss} && mocha test/**/*.js"
+    "lint": "eslint 'test/**/*.js' && stylelint '{_index,test/**/*}.scss'",
+    "postpublish": "GITHUB_TOKEN=$GITHUB_RELEASE_TOKEN github-release-from-changelog",
+    "release": "np --no-release-draft",
+    "test": "npm run lint && mocha --require esm 'test/**/*.js'",
+    "test:watch": "npm test -- --watch",
+    "version": "version-changelog CHANGELOG.md && changelog-verify CHANGELOG.md && git add CHANGELOG.md"
   },
   "devDependencies": {
-    "eslint": "^3.1.1",
-    "eslint-config-niksy": "^3.0.1",
-    "mocha": "^2.5.3",
-    "sass-true": "^2.1.3",
-    "stylelint": "^7.0.3",
-    "stylelint-config-niksy": "^3.0.1"
+    "changelog-verify": "^1.1.2",
+    "eslint": "^7.11.0",
+    "eslint-config-niksy": "^9.0.0",
+    "eslint-config-prettier": "^6.14.0",
+    "eslint-plugin-extend": "^0.1.1",
+    "eslint-plugin-import": "^2.22.1",
+    "eslint-plugin-jsdoc": "^30.7.3",
+    "eslint-plugin-mocha": "^8.0.0",
+    "eslint-plugin-node": "^11.1.0",
+    "eslint-plugin-prettier": "^3.0.1",
+    "eslint-plugin-promise": "^4.1.1",
+    "eslint-plugin-unicorn": "^23.0.0",
+    "esm": "^3.0.51",
+    "github-release-from-changelog": "^2.1.1",
+    "husky": "^4.3.0",
+    "lint-staged": "^10.4.2",
+    "mocha": "^8.2.0",
+    "np": "^6.5.0",
+    "prettier": "^2.1.2",
+    "sass": "^1.32.4",
+    "sass-true": "^6.0.1",
+    "stylelint": "^13.7.2",
+    "stylelint-config-niksy": "^8.0.0",
+    "stylelint-config-prettier": "^8.0.2",
+    "stylelint-prettier": "^1.1.2",
+    "version-changelog": "^3.1.1"
   },
   "engines": {
-    "node": ">=4"
+    "node": ">=10"
   },
   "keywords": [
     "element",
